requirements.txt
BeautifulSoup4
lxml
PyYAML
twitter
